refactor(esm): register the analyzable naming pass from its condition
The deferred fill was a plugin class two other plugins each applied
defensively, kept single by a module-level `WeakSet` of compilers. Nothing
reserves a stand-in without `output.module`, so register it there instead:
once, from the condition that decides it.

The pass cannot arrange this itself from the reservation — a module restored
from the cache carries a stand-in without being generated again.

`AsyncWebAssemblyModulesPlugin` applied it unconditionally, so a non-ESM async
wasm build scanned every emitted asset for stand-ins it could not contain.
