+++ lib/RuntimeTemplate.js
@@ -84,8 +84,8 @@ const CHUNK_ASSET_NAMING = new Map([
 	]
 ]);
 const getTemplatedPathPlugin = memoize(() => require("./TemplatedPathPlugin"));
-const getAnalyzableChunkHashPlugin = memoize(() =>
-	require("./esm/AnalyzableChunkHashPlugin")
+const getAnalyzableChunkNaming = memoize(() =>
+	require("./esm/analyzableChunkNaming")
 );
 const getConcatenatedModule = memoize(() =>
 	require("./optimize/ConcatenatedModule")
@@ -203,7 +203,7 @@ const SERVED_BAILOUT =
  * @import {
  * 	SpecifierPart,
  * 	SpecifierPartKind
- * } from "./esm/AnalyzableChunkHashPlugin"
+ * } from "./esm/analyzableChunkNaming"
  */
 /** @typedef {"import" | "url" | "url-runtime" | "url-inline" | "wasm" | "wasm-relative"} AnalyzableForm */
 /** @import AsyncDependenciesBlock from "./AsyncDependenciesBlock" */
@@ -840,7 +840,7 @@ class RuntimeTemplate {
 			// opening with a letter reads as a scheme and drops the base entirely.
 			const text =
 				spelled !== null &&
-				!getAnalyzableChunkHashPlugin().hasReservedFullHash(spelled)
+				!getAnalyzableChunkNaming().hasReservedFullHash(spelled)
 					? spelled
 					: null;
 			/** @type {URL} */
@@ -862,7 +862,7 @@ class RuntimeTemplate {
 				return null;
 			}
 			return toJsStringLiteral(
-				getAnalyzableChunkHashPlugin().reserveSpecifier([
+				getAnalyzableChunkNaming().reserveSpecifier([
 					["base", base],
 					...parts,
 					["literal", filename]
@@ -2846,7 +2846,7 @@ class RuntimeTemplate {
 			}
 			if (!canReserve) return cannotReserve();
 			return toJsStringLiteral(
-				getAnalyzableChunkHashPlugin().reserveSpecifier([
+				getAnalyzableChunkNaming().reserveSpecifier([
 					...prefix,
 					[naming.standIn, /** @type {ChunkId} */ (chunk.id)]
 				])
@@ -2964,8 +2964,8 @@ class RuntimeTemplate {
 			this._analyzableBailout(module, DEFER_BAILOUT);
 			return null;
 		}
-		const analyzableChunkHashPlugin = getAnalyzableChunkHashPlugin();
-		if (!analyzableChunkHashPlugin.canDeferFullHash(publicPath)) {
+		const analyzableChunkNaming = getAnalyzableChunkNaming();
+		if (!analyzableChunkNaming.canDeferFullHash(publicPath)) {
 			return [["template", publicPath]];
 		}
 		return [
@@ -2973,7 +2973,7 @@ class RuntimeTemplate {
 				"literal",
 				this.compilation.getPath(
 					publicPath,
-					analyzableChunkHashPlugin.DEFERRED_FULL_HASH_PATH_DATA
+					analyzableChunkNaming.DEFERRED_FULL_HASH_PATH_DATA
 				)
 			]
 		];
@@ -3065,7 +3065,7 @@ class RuntimeTemplate {
 				return null;
 			}
 			return toJsStringLiteral(
-				getAnalyzableChunkHashPlugin().reserveSpecifier(parts)
+				getAnalyzableChunkNaming().reserveSpecifier(parts)
 			);
 		};
 		if (bakePublicPath && publicPath !== "auto") {
@@ -3271,7 +3271,7 @@ class RuntimeTemplate {
 					module,
 					runtime,
 					chunkGraph,
-					...getAnalyzableChunkHashPlugin().DEFERRED_FULL_HASH_PATH_DATA
+					...getAnalyzableChunkNaming().DEFERRED_FULL_HASH_PATH_DATA
 				})
 			)}`
 		);
