+++ lib/WebpackOptionsApply.js
@@ -522,6 +522,13 @@ class WebpackOptionsApply extends OptionsApply {
 			}
 		}
 
+		if (options.output.module) {
+			// Analyzable output is ESM output, so this is every build that can reserve a name.
+			const { fillReservedNames } = require("./esm/analyzableChunkNaming");
+
+			fillReservedNames(compiler);
+		}
+
 		if (options.experiments.syncWebAssembly) {
 			const WebAssemblyModulesPlugin = require("./wasm-sync/WebAssemblyModulesPlugin");
