+++ lib/esm/ModuleChunkLoadingPlugin.js
@@ -7,7 +7,6 @@
 
 const RuntimeGlobals = require("../RuntimeGlobals");
 const lazyModule = require("../util/lazyModule");
-const AnalyzableChunkHashPlugin = require("./AnalyzableChunkHashPlugin");
 
 /** @import Chunk from "../Chunk" */
 /** @import Compiler from "../Compiler" */
@@ -30,8 +29,6 @@ class ModuleChunkLoadingPlugin {
 	 * @returns {void}
 	 */
 	apply(compiler) {
-		// Fills in the specifiers this plugin's analyzable `import()` reserves.
-		new AnalyzableChunkHashPlugin().apply(compiler);
 		compiler.hooks.thisCompilation.tap(PLUGIN_NAME, (compilation) => {
 			const globalChunkLoading = compilation.outputOptions.chunkLoading;
 			/**
