+++ lib/wasm-async/AsyncWebAssemblyModulesPlugin.js
@@ -10,7 +10,6 @@ const Generator = require("../Generator");
 const { WEBASSEMBLY_MODULE_TYPE_ASYNC } = require("../ModuleTypeConstants");
 const WebAssemblyImportDependency = require("../dependencies/WebAssemblyImportDependency");
 const { tryRunOrWebpackError } = require("../errors/HookWebpackError");
-const AnalyzableChunkHashPlugin = require("../esm/AnalyzableChunkHashPlugin");
 const { compareModulesByFullName } = require("../util/comparators");
 const createHooksRegistry = require("../util/createHooksRegistry");
 const lazyModule = require("../util/lazyModule");
@@ -82,7 +81,6 @@ class AsyncWebAssemblyModulesPlugin {
 		// A binary's name may carry the compilation hash, which the analyzable call site
 		// leaves as a stand-in — so whatever the chunk loading is, the pass that fills it
 		// in has to be there. Applying it twice is a no-op.
-		new AnalyzableChunkHashPlugin().apply(compiler);
 		compiler.hooks.compilation.tap(
 			PLUGIN_NAME,
 			(compilation, { normalModuleFactory }) => {
