+++ test/configCases/analyzable/stand-in-lookalike/lookalikes.js
@@ -1,6 +1,6 @@
 // Text of our own that spells the stand-in the deferred pass fills in. Nothing about
 // a payload is given, so each of these has to reach the bundle exactly as written.
-// Held as json so the unit test can state what each one decodes to.
+// Held as json so the case can state what each one decodes to.
 import lookalikes from "./lookalikes.json";
 
 export { lookalikes };
